requirements.txt
torch>=2.0.0
torchvision>=0.15.0

# SAT 框架 (SwissArmyTransformer)
SwissArmyTransformer>=0.4.4

transformers>=4.27.1

bitsandbytes>=0.41.0

Pillow>=9.0.0

numpy>=1.24.0
requests>=2.28.0
tqdm>=4.65.0

# 分布式训练相关 
deepspeed>=0.9.0

tensorboardX>=2.6

datasets>=2.12.0

cpm_kernels>=1.0.11
sentencepiece>=0.1.99

einops>=0.6.1
mdtex2html>=1.2.0
